pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["radosz99  <241385@student.pwr.edu.pl>"]

[tool.poetry.dependencies]
python = "^3.8"
asgiref = "3.5.2"
"backports.zoneinfo" = "0.2.1"
Django = "4.1"
sqlparse = "0.4.2"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
